Adjust badges
The Travis badge was not pointing to the master branch so any failure on any branch would make the badge in the readme show a build failure. The Codacy badge from shields.io is currently broken https://github.com/badges/shields/issues/1985 I have commented it to hide for now as well as pointed it to the master branch for when it is fixed.