Partial fix to multiple instances of a variable in .env
A partial fix to having multiple instances of a variable in the `.env`.  Previously, it would concatenate the values, so that

```
VAR='one'
VAR='two'
```

becomes

'''
VAR='one two'
```

This partially fixes it by only finding the first line in `env_set`, and only finding the last line in `env_get`.  The full fix will be to be sure the `.env` file isn't sorted beforehand, so that the last entry in the file will be the value used.