Implement an optional `Override` argument to `question_prompt`
Make `question_prompt` take an optional `Override` parameter at the end (currently, the fourth parameter).  If set to `Y` or `N`, make `question_prompt` immediately return as if the user entered that selection.

This is used by adding `"${VERBOSE:+Y}"` or `"${FORCE:+Y}"` at the end of the function calls, only adding `Y` if either of those is set and not empty, forcing that return value.