Update main.sh
Create a `highlight_list` function.

It takes a whitespace/newline delimited set of words, and outputs it as whitespace delimited list.  Each word is surrounded by `\Zr` and `\ZR`, to reverse the colors of each word in `dialog` with the `--colors` option.

This function is used to set the subtitle of calls to `run_script_dialog` when there are multiple values passed to a function (such as multiple apps being added).