Update main.sh
Change `--env-get`, `--env-get-lower`, `--env-set` and `--env-set-lower` to allow the syntax of `--env-get var1 var2` and `--env-set var=val`.

Still allow the previous syntax of `--env-get=var` and `--env-set=var,val` for anyone that may be using them in scripts.

Rename `--env-lines` to `--env-appvar-lines` and `--env-list` to `--env-appvar-list`.

Allow the use of the `-g/--gui` option and the use of multiple variables and app names in the `--env-get`, `--env-appvar-lines` and `--env-appvar-list` options.

Update the usage text with the new syntax, including where multiple values are allowed.