Update menu_add_var.sh
Allow to add a new variable.  The application the variable is for is passed as the only argument, witch can be a blank string for global variable in the `.env1 file, `appname` for an `APPNAME__*` variable in the ``.env` file, or `appname:` for a variable in `env_files/appname.env`.

It will not give an error if the variable already exists, or if the veriable is not global or for the right app, depending on the appname passed.