Set global ROWS and COLS variables
Make global `ROWS` and `COLS` variables instead of using the system `LINES` and `COLUMNS` variables.  Make the base `dialog_*` command update the values, as well as `menu_heading`.  For menus that use the heading from `menu_heading`, try to call that routine just before the value is used in case the terminal size has been changed since the last call.