Add global defaults for turning on STORAGE mounts
Add global defaults for the mounting of the STORAGE volumes.  The logic is now:

`DOCKER_MULTIPLE_STORAGE='ON'` turns multiple storage folders on or off.  If it is not set to ON, YES or TRUE (ignoring case), only the main STORAGE folder will be mounted in containers.

`DOCKER_VOLUME_STORAGE`, `DOCKER_VOLUME_STORAGE2`, `DOCKER_VOLUME_STORAGE3`, and `DOCKER_VOLUME_STORAGE4` are the four variables containing the folders to mount.  If the variable is set to an empty string, the mount is not done.

`DOCKER_STORAGE`, `DOCKER_STORAGE2`, `DOCKER_STORAGE3` and `DOCKER_STORAGE4` are boolean variables to set the default mount condition of each STORAGE folder.  If they are set to ON, TRUE or YES, the default is to mount them.  Mount folders 2-4 are still not mounted unless the global boolean `DOCKER_MULTIPLE_STORAGE` is on.

Each app has it's own set of boolean `APPNAME__STORAGE`, `APPNAME__STORAGE2`, `APPNAME__STORAGE3`  and `APPNAME__STORAGE4`  variables. If one of these is not an empty string, it uses it's value instead of the matching global variable to determine if the mount for that STORAGE folder is done. Again, `DOCKER_MULTIPLE_STORAGE` needs to be turned on for mount folders 2-4 to apply.