Add code to switch from branch `master` to `main`
At the start of each run, check if the current branch is named `master`, and if a branch named `main` exists.  If it does, force an update to branch `main`, then re-run the script with the command-line options given.

Portions of `main.sh` had to be re-arranged so variables were already defined for the earlier call to `update_self`, and checking for dialog box usage.

The command-line arguments are also passed to `update_self` after the optional branch name, with the command to run after updates.  If not supplied, it uses the previous default of `bash ${SCRIPTNAME} -e`