Create `app_is_runnable`, `app_filter_runnable` and `app_filter_runnable_pipe`
Create `app_is_runnable` to return a true value if the app name passed to the function is runnable on the current system.

Create `app_filter_runnable` to output only the app names of the apps that are runnable, from the list of app names passed to the function.

Create `app_filter_runnable_pipe` to operate the same as `app_filter_runnable`, but taking its arguments from `stdin`, to be used in a pipe.

Use `app_filter_runnable_pipe` in `menu_app_select` to speed up the generation of the app list.