Re-work `menu_app_select` getting rid of temporary file usage
Display a static info box during generation of the application list.
Overwrite the info box with a new one after detecting the selected applications.
Suppress any possible notices output to `stderr` while the info box is shown.
Add to the `AppsLIst` array directly instead of using the temporary files that were required do to doing processing in a subprocess created by a pipe.