Re-write of the progress bar in `menu_app_select`
Re-write the the display of the progress bar in `menu_app_select`.

Remove the usage of multiple calls to `dialog --mixedgauge`, and instead use a single background process of `dialog --gauge`, passing the updated status percent and messages through a named pipe.  This gets rid of the flashing on the screen for each call to `dialog --mixedgauge`, and allows for more flexibility in the status messages.