Check if the original app template file has changed before create an instance file (#2123)
Remove the clearing of the instances folder in `update_self`.
In `app_instance_file`, check for file `.instances/.apps/baseappname/baseappname.extension` .  If it doesn't exist, or is not the same as the original file in `.apps/baseappname/baseappname.extension`, then copy the original file there, and create the new instance file in `.instances/appname/appname.extension`.

This loses a bit of time in general to compare files, but avoids having to recreate all instance files after an update.