Look in override file to check for referenced apps
Search the override for for `.env.app.appname` strings, and if found, consider `appname` to be referenced.

Delete any `.env.app.appname` files for apps that are no longer referenced, i.e., do not have any `APPNAME__*` variables in the global `.env` file, do not have any variables in the `.env.app.appname` file, and do not have `.env.app.appname` in the override file.