Update the default theme to be more usable on terminals that use a dark blue color
Update the default DockSTARTer theme, using CYAN as the main color, instead of BLUE, to be more usable in terminals that use a dark blue color for BLUE.  Adjust other colors for readability on the new color background.

Rename the old DockSTARTer theme to OldDockerSTARTer to still be usable for those that prefer it.